Don't construct `report_issue_url` through string manipulation
Instead of constructing the URL for reporting an issue by manipulating
the raw string, we can instead use the `URI.encode_www_form`[0] method.

This ensures that we're constructing the URL as a well-formed, encoded
set of parameters, and has the added benefit of it being much clearer to
read, too.

[0]: https://stackoverflow.com/a/11251654