Replace sidebar with nav bar at top of page
We felt the sidebar was a bit busy on the page, and thought having a simpler
navbar at the top wouldn't make it any harder to navigate the site.

I've reused the styling from the contribution bar at the bottom of the page,
which comes from the tech docs template (https://tdt-documentation.london.cloudapps.digital/configure_project/global_configuration/#show-contribution-banner)
but added some extra margin so the text isn't right next to the search bar.