WIP: Use consistent schema for standards' pages
As a step towards being able to produce metadata for the standards
pages, we should use a consistent model across our standards.

To do this, we add a JSON schema for the standards, which attempts to
make as much required and formatted appropriately.

This requires we:

- wrap string fields, where appropriate, in string quotes
- remove the redundant `name` field, as it duplicates `title`
- remove the redundant `category` field, as it duplicates
  `classification`
- where fields are likely to be not free-form, we can mark them as
  `enum`s

This uses the `json-schema` library, as it's got nice error reporting,
but unfortunately is stuck on `draft-04` of JSON Schema.