Remove /standards-catalogue/ prefix to site
As noted in #90, the additional /standards-catalogue/ hierarchy,
especially if going to /standards/, can be quite awkward to use.

We can remove this by amending our Tech Docs + Middleman configuration
to no longer use a sub-project from the top-level repo, so this site is
for the DSA project.

Now we're no longer needing the subdirectory, we can remove the
HTML-only redirect.

Closes #90.