Fix: Remove erroneous URL portions in feedback links
As spotted in #102, the URLs that are being referenced from the
"Feedback" links are not correct, as they've got an additional
`/data-standards-authority` in them.

A short-term fix is to trim the first instance of the prefix.

It appears that this is an artefact of something we've done on our
mix of Tech Docs template and custom code, as other Tech Docs
projects[0] running on GitHub pages seem to work OK.

Closes #102.

[0]: https://alphagov.github.io/accessibility-personas/ashleigh/