Add JSON Schema for Standards
Now we have an agreed data model for the Standards portion of the
Workbench, we should codify it with a JSON schema.

This uses the `json-schema` library, as it's got nice error reporting,
but unfortunately is stuck on `draft-04` of JSON Schema.

Note that although this adds it to the Rake tasks, it is not applied in
CI, so PRs will continue to succeed to build until the schemas are
updated.