Validate keys in data files
To make sure that the keys used for data files stay aligned with the
expected naming standards, we should make sure that we add logic to
validate these.

As we can do this with a single regex (for now) we can add this to the
`is_data_file_schema_compliant` method, and report any issues to the
user, with a human-readable error for what has failed, and why.

Closes #137.