Make `contraction` an optional property for standards
As noted in ADR-0003 and as part of #135, we want to make this optional.

However, when it's present, we want to make sure that it's non-empty, so
need to add a `minLength` check.

As it's optional, we need to make sure we're not populating it in our
schema.org metadata, nor is it required to be output into the tabular
information for the standard.