Split build and publish functionality into jobs
Because these are two independent pieces of work, we can split them into
different jobs.

This allows us to split out the logic between the build and deploy, as
well as allowing us to look at tweaking permissions for our jobs in
future commits.

We need to upload/download the built site between jobs using
`upload-artifact` / `download-artifact`.

We also need to make sure that the built site is downloaded to `build`,
as that's where our Rake task expects it to be.