Include short description, tags and date updated in contents
This should make it clearer when people land on this page what the standards
are about and make it easier to find one that fits their needs.

If `short_description` is not provided in the front matter, it will just
display the name on the contents page rather than name and description.

`short_description` has been added as a property to our standards schema, and
at the moment can be null, but we'll make this mandatory when we have some
descriptions added for the existing standards.